+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All Notable changes to `pdf-to-image` will be documented in this file
 
+## 1.2.0 - 2016-04-29
+- added `saveAllPagesAsImages`-function.
+
+## 1.1.0 - 2015-04-13
+- added `getImageData`-function.
+
 ## 1.0.3 - 2015-01-22
 
 ### Bugfix